+++ src/main/java/com/CuteNekoDragon/Core/common/CommonProxy.java
@@ -1,10 +1,14 @@
 package com.CuteNekoDragon.Core.common;
 
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 import com.CuteNekoDragon.Core.SVOCore;
 import com.CuteNekoDragon.Core.common.data.SVOCreativeTab;
+import com.CuteNekoDragon.Core.common.data.SVOItemPrices;
+import com.CuteNekoDragon.Core.common.data.SVOItemTooltips;
 import com.CuteNekoDragon.Core.common.data.SVOItems;
 import com.CuteNekoDragon.Core.config.SVOConfig;
 
@@ -20,5 +24,13 @@ public class CommonProxy {
 
         SVOItems.init();
         SVOCreativeTab.init();
+    }
+
+    @SubscribeEvent
+    public void onCommonSetup(final FMLCommonSetupEvent event) {
+        event.enqueueWork(() -> {
+            SVOItemPrices.ProvidePrices();
+            SVOItemTooltips.ProvideTooltips();
+        });
     }
 }