+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOItemPrices.java
@@ -1,0 +1,10 @@
+package com.CuteNekoDragon.Core.common.data;
+
+import com.CuteNekoDragon.Core.utils.PriceUtil;
+
+public class SVOItemPrices {
+
+    public static void ProvidePrices() {
+        PriceUtil.setPrice(SVOItems.QUARTZ, 25);
+    }
+}