+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOItemTooltips.java
@@ -1,0 +1,11 @@
+package com.CuteNekoDragon.Core.common.data;
+
+import com.CuteNekoDragon.Core.utils.TooltipBuilder;
+
+public class SVOItemTooltips {
+
+    public static void ProvideTooltips() {
+        TooltipBuilder.addTooltip(SVOItems.QUARTZ).addCoins().addGlyph(TooltipBuilder.SVOTypes.MINERAL)
+                .addInfo("tooltip.svo_core.quartz");
+    }
+}